refactor AddressFamily
Move it to common package in order to avoid import cycle.
